+++ Sources/Rapid/Equatable/EquatableRange.swift
@@ -1,0 +1,33 @@
+// EquatableRange.swift
+// Copyright © 2023 Kaleb A. Ascevich
+//
+// This package is free software: you can redistribute it and/or modify it
+// under the terms of the GNU General Public License as published by the
+// Free Software Foundation, either version 3 of the License, or (at your
+// option) any later version.
+//
+// This package is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+// FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+// for more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this package. If not, see https://www.gnu.org/licenses/.
+
+extension PartialRangeFrom: Equatable {
+    public static func == (lhs: Self, rhs: Self) -> Bool {
+        lhs.lowerBound == rhs.lowerBound
+    }
+}
+
+extension PartialRangeThrough: Equatable {
+    public static func == (lhs: Self, rhs: Self) -> Bool {
+        lhs.upperBound == rhs.upperBound
+    }
+}
+
+extension PartialRangeUpTo: Equatable {
+    public static func == (lhs: Self, rhs: Self) -> Bool {
+        lhs.upperBound == rhs.upperBound
+    }
+}